requirements.txt
Django
psycopg2
gunicorn
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
djangorestframework-jwt
python-dateutil
django-crontab
django-admin-rangefilter